==> notebooks/4. Holiday.ipynb <==
# 대한민국 공휴일 데이터
holiday = 신정, 설, 추석
anniversary = 위 제외한 대한민국 공휴일

pip install holidays

from datetime import date
import holidays
import pandas as pd
import numpy as np

kr_holidays = holidays.KR()

tmp = []

for date, name in sorted(holidays.KR(years=[2018,2019,2020,2021,2022,2023]).items()):
    print(date, name)
    tmp.append([date, name])

data = pd.DataFrame(data = tmp, columns=['배송년월일', '공휴일'])

data['배송년월일'] = pd.to_datetime(data['배송년월일']).dt.strftime('%Y%m%d')

data['공휴일'].value_counts()

dates = pd.date_range(start='2018-01-01', end='2023-12-31', freq='D')
dates_df = pd.DataFrame({'배송년월일': dates.strftime('%Y%m%d')})
dates_df

holiday_df = pd.merge(dates_df, data, on='배송년월일', how='left')

def 명절(x):
    if '설날' in x or '추석' in x or '신정' in x:
        return 1
    else:
        return 0

def 기념일(x):
    if '설날' in x or '추석' in x or '신정' in x:
        return 0
    elif x:
        return 1
    else:
        return 0

holiday_df['공휴일'].fillna('', inplace=True)
holiday_df['holiday'] = holiday_df['공휴일'].apply(명절)
holiday_df['anniversary'] = holiday_df['공휴일'].apply(기념일)

holiday_df.tail(10)

holiday_df = holiday_df[['배송년월일', 'holiday', 'anniversary']]
holiday_df.rename(columns={'배송년월일':'date'}, inplace=True)

holiday_df.to_csv("holiday.csv", index=False, encoding='euc-kr')
